Shuffle CLI commands
