+++ choreographer/__init__.py
@@ -1,4 +1,5 @@
 from .browser import Browser
 from .browser import which_browser
+from .cli_utils import get_browser
 
-__all__ = [Browser, which_browser]
+__all__ = [Browser, which_browser, get_browser]
