+++ pyproject.toml
@@ -35,7 +35,8 @@ dev = [
 ]
 
 [project.scripts]
-dtdoctor = "choreographer.browser:diagnose"
+choreo_diagnose = "choreographer.cli_utils:diagnose"
+choreo_get_browser = "choreographer.cli_utils:get_browser"
 
 [tool.ruff.lint]
 ignore = ["E701"]
